global extract

+++ src/vdbvoxelgrid/pybind/vdbvoxelgrid_pybind.cpp
@@ -150,7 +150,7 @@ PYBIND11_MODULE(vdbvoxelgrid_pybind, m) {
                 // this is not necessary, pybind11 can do it, but this is 5 times faster
                 pybind11::dict ret_dict;
                 for (auto it = map.begin(); it != map.end(); it++) {
-                    ret_dict[py::str(it->first)] = py::array_t<int>(it->second.size(), it->second.data());
+                    ret_dict[py::str(it->first)] = py::array_t<float>(it->second.size(), it->second.data());
                 }
                 return ret_dict;
             });
